Cleaned repository

+++ cpuencode.h
@@ -4,7 +4,7 @@
 extern "C"
 void cpu_vlc_encode(unsigned int* indata, unsigned int num_elements, 
 					unsigned int* outdata, unsigned int *outsize, 
-					uint256 *codewords, unsigned int* codewordlens);  
+					unsigned int *codewords, unsigned int* codewordlens);  
 #endif
